ai-engine/requirements.txt
fastapi
uvicorn
groq
pydantic
python-multipart
requests
python-dotenv

langchain
langchain-community
chromadb
sentence-transformers
tiktoken